Adding spans for setup, teardown, and individual fixtures
In test suites with complex setups, teardowns, and fixtures, it's common to see
most of the test runtime happening in those stages rather than in the individual
tests themselves.

In this change, session- and module-scoped fixtures are attributed to the
overall test session, while function-scoped fixtures are attributed to the
setup for an individual test.  This will help with artificial skew from tests
that happen to be the first ones that request a higher-scoped fixture.

I'd welcome feedback on the layout of the spans after some folks have had time
to try it out.  `pytest` doesn't have a clearly defined session or module
`setup` stage where these fixtures are run, because they are invoked lazily
the first time a test requests them.  This is ideal for performance, but these
spans will end up kind of "hanging in mid-air" between other test suites.
Depending on the visualization tool (OpenObserve, Jaeger, etc), you may see
those higher-scoped fixture spans showing up in different spots.

Thanks to @drcraig for the idea and @sihil for the cheers!

Closes #25